requirements.txt
colorama==0.4.4
commonmark==0.9.1
cycler==0.10.0
Faker==8.11.0
kiwisolver==1.3.1
matplotlib==3.4.2
networkx==2.6.2
numpy==1.21.1
Pillow==8.3.1
Pygments==2.7.4
pyparsing==2.4.7
python-dateutil==2.8.1
rich==10.7.0
six==1.15.0
text-unidecode==1.3
typing-extensions==3.7.4.3
python-telegram-bot==13.7